Lootable Chest & Parry
Implemented chests with random loot inside and open animations. Also implemented a parry system with a riposte attack inflicting critical damage. Also implemented all the proper animations and improved the stamina system so that it works properly with all the player movements and prevents him to do certain activities while it's completely drained.

+++ Assets/Systems/AI/Scripts/PlayerStats.cs
@@ -93,7 +93,7 @@ public class PlayerStats : CharacterStats
         }
     }
 
-    public void TakeStaminaDamage(int damage)
+    public void TakeStaminaDamage(float damage)
     {
         currenStamina = currenStamina - damage;
         staminaBar.SetCurrentStamina(currenStamina);
